Added Dynamic Pivoting Store Procedure
The following is a store procedure that takes input such as table name, pivot column, aggregation and non-aggregated column and outputs as a pivot table. The code is intended to work only on MySQL and MariaDB since there is no inbuilt PIVOT function, but can be tweaked to accommodate any SQL flavors.